+++ redist/rom/game_objects/game_object_flipper.h
@@ -22,6 +22,6 @@ namespace spintool::rom
 
 		constexpr static const size_t s_size_on_rom = 0xA;
 
-		static FlipperInstance LoadFromROM(const rom::SpinballROM& rom, size_t offset);
+		static FlipperInstance LoadFromROM(const rom::SpinballROM& rom, Uint32 offset);
 	};
 }
