+++ redist/rom/rom_data.h
@@ -6,12 +6,12 @@ namespace spintool::rom
 {
 	struct ROMData
 	{
-		size_t rom_offset = 0;
-		size_t real_size = 0;
-		size_t rom_offset_end = 0;
+		Uint32 rom_offset = 0;
+		Uint32 real_size = 0;
+		Uint32 rom_offset_end = 0;
 
-		void SetROMData(const size_t start_offset, size_t end_offset);
-		void SetROMData(const Uint8* start_data_ptr, const Uint8* end_data_ptr, const size_t start_offset);
+		void SetROMData(const Uint32 start_offset, Uint32 end_offset);
+		void SetROMData(const Uint8* start_data_ptr, const Uint8* end_data_ptr, const Uint32 start_offset);
 
 		bool operator==(const ROMData& rhs) const
 		{
