+++ redist/rom/ssc_decompressor.h
@@ -15,8 +15,8 @@ namespace spintool::rom
 	class SSCDecompressor
 	{
 	public:
-		static SSCDecompressionResult DecompressData(const std::vector<Uint8>& in_data, size_t offset, size_t working_data_size_hint);
-		static SSCDecompressionResult IsValidSSCCompressedData(const Uint8* in_data, const size_t starting_offset);
+		static SSCDecompressionResult DecompressData(const std::vector<Uint8>& in_data, Uint32 offset, Uint32 working_data_size_hint);
+		static SSCDecompressionResult IsValidSSCCompressedData(const Uint8* in_data, const Uint32 starting_offset);
 	private:
 	};
 }
