+++ redist/rom/tileset.h
@@ -30,16 +30,16 @@ namespace spintool::rom
 	struct TileSet
 	{
 		ROMData rom_data;
-		size_t compressed_size = 0;
-		size_t uncompressed_size = 0;
+		Uint32 compressed_size = 0;
+		Uint32 uncompressed_size = 0;
 
 		Uint16 num_tiles = 0;
 		std::vector<Uint8> uncompressed_data;
 
-		static TilesetEntry LoadFromROM(const SpinballROM& src_rom, size_t rom_offset);
-		static TilesetEntry LoadFromROMSecondCompression(const SpinballROM& src_rom, size_t rom_offset);
-		std::shared_ptr<const Sprite> CreateSpriteFromTile(const size_t offset) const;
-		std::shared_ptr<SpriteTile> CreateSpriteTileFromTile(const size_t tile_index) const;
+		static TilesetEntry LoadFromROM(const SpinballROM& src_rom, Uint32 rom_offset);
+		static TilesetEntry LoadFromROMSecondCompression(const SpinballROM& src_rom, Uint32 rom_offset);
+		std::shared_ptr<const Sprite> CreateSpriteFromTile(const Uint32 offset) const;
+		std::shared_ptr<SpriteTile> CreateSpriteTileFromTile(const Uint32 tile_index) const;
 		constexpr const static Uint16 s_tile_width = 0x08;
 		constexpr const static Uint16 s_tile_height = 0x08;
 		constexpr const static Uint16 s_tile_total_pixels = s_tile_width * s_tile_height;
